Using build-specific files again to fix windows build
