+++ go/service/ctl.go
@@ -7,8 +7,6 @@ import (
 	"golang.org/x/net/context"
 
 	"github.com/keybase/client/go/engine"
-	"github.com/keybase/client/go/install"
-	"github.com/keybase/client/go/launchd"
 	"github.com/keybase/client/go/libkb"
 	keybase1 "github.com/keybase/client/go/protocol"
 	rpc "github.com/keybase/go-framed-msgpack-rpc"
@@ -29,30 +27,12 @@ func NewCtlHandler(xp rpc.Transporter, v *Service, g *libkb.GlobalContext) *CtlH
 }
 
 // Stop is called on the rpc keybase.1.ctl.stop, which shuts down the service.
+//
+// stop() is defined in ctl_osx.go and ctl_non_osx.go for
+// OS-specific functionality.
 func (c *CtlHandler) Stop(_ context.Context, args keybase1.StopArg) error {
 	c.G().Log.Debug("Received stop(%d) RPC; shutting down", args.ExitCode)
-
-	if c.service.ForkType == keybase1.ForkType_LAUNCHD {
-		c.stopLaunchd()
-	}
-
-	go c.service.Stop(args.ExitCode)
-
-	return nil
-}
-
-func (c *CtlHandler) stopLaunchd() {
-	status := install.KeybaseServiceStatus(c.G(), c.G().Env.GetLabel())
-	if status.Pid == "" {
-		c.G().Log.Debug("Service does not appear to be running via launchd (label = %q)", c.G().Env.GetLabel())
-		return
-	}
-
-	c.G().Log.Debug("Removing %s from launchd", status.Label)
-	svc := launchd.NewService(status.Label)
-	if err := svc.Stop(false); err != nil {
-		c.G().Log.Warning("error stopping launchd service:", err)
-	}
+	return c.stop(args)
 }
 
 func (c *CtlHandler) LogRotate(_ context.Context, sessionID int) error {
